Use Debian instead of Alpine Linux
The project uses Alpine now, I assume the reason why is to have
a more resource efficient Docker container.

This fixes the issue with Chrome not starting properly when using Karma.
Alpine Linux is great for small Docker images, buzzword friendly
and all, but tends to be quite poorly supported by Google Chrome.
(It's a hack at best. Running Google Chrome inside a Docker container
is quite a stretch in the first place, it's not exactly how Chrome was
designed to run...)

Although it seems to be possible to run Chrome in Alpine as well,
that method introduces its own bunch of problems, and when it's really
acting up, consider that the time a person spends debugging stuff is
much more valuable than the time and CPU Alpine Linux will save you
compared to a Debian-based image.